+++ raster/src/test/scala/geotrellis/raster/io/geotiff/BitGeoTiffTileSpec.scala
@@ -17,12 +17,6 @@
 package geotrellis.raster.io.geotiff
 
 import geotrellis.raster._
-import geotrellis.raster.io.geotiff.writer.GeoTiffWriter
-import geotrellis.raster.mapalgebra.local._
-
-import geotrellis.vector.Extent
-
-import geotrellis.proj4._
 
 import geotrellis.raster.testkit._
 
@@ -81,6 +75,23 @@ class BitGeoTiffTileSpec extends FunSpec
       val res = tile.mapDouble { (col, row, z) => z }
 
       assertEqual(res, tile)
+    }
+
+    it("should convert tiled tiffs") {
+      val tiff = SinglebandGeoTiff(geoTiffPath("bilevel.tif"))
+      val tile = tiff.tile.toArrayTile()
+
+      // check that it is possible to convert bit cellType to bit cellType
+      val tiffTile = tile.toGeoTiffTile.convert(BitCellType)
+      assertEqual(tiffTile.toArrayTile(), tile.toArrayTile)
+
+      // check that it is possible to convert int cellType to bit cellType
+      // and that bitCellType conversion is idempotent
+      (0 to 5).foldLeft(tile.toGeoTiffTile.convert(IntCellType)) { case (acc, _) =>
+        val tiffTileLocal = acc.convert(BitCellType)
+        assertEqual(tiffTileLocal.toArrayTile(), tile.toArrayTile)
+        tiffTileLocal
+      }
     }
   }
 }